tests/wpt/tests/content-security-policy/resource-hints/prefetch-no-csp.html

<!DOCTYPE html>
<html>
<head>
  <script src='/resources/testharness.js'></script>
  <script src='/resources/testharnessreport.js'></script>
  <script src='/common/utils.js'></script>
  <script src='/content-security-policy/support/testharness-helper.js'></script>
  <script src='/content-security-policy/support/prefetch-helper.js'></script>
  <script>
    promise_test(async t => {
      assert_true(await try_to_prefetch("/common/dummy.xml", t));
    }, 'Prefetch succeeds when no CSP');
  </script>
</head>
<body>
</body>
</html>
